Revert "Feature/cu 18 ssr (#91)"
This reverts commit 0553d642400ab5f1bdfc22c54831a46465072378.
