+++ src/app/+state/product.store.ts
@@ -1,9 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Observable, catchError, combineLatest, forkJoin, of, tap } from 'rxjs';
-import { SignalService } from '../shared/services/state.service';
-import { ExternalProduct } from '../shared/interfaces/products.interface';
-import { ProductRepository } from '../shared/services/repository/product.repository';
-import { LoaderService } from '../shared/components/loader/service/loader.service';
+import { ExternalProduct } from 'src/app/shared/interfaces/products.interface';
+import { ProductRepository } from 'src/app/shared/services/repository/product.repository';
+import { SignalService } from 'src/app/shared/services/state.service';
 
 export const initialState: initialProductState = {
   externalProducts: [],
@@ -22,17 +21,13 @@ export class ProductState extends SignalService<initialProductState> {
     return this.state.externalProducts;
   }
 
-  constructor(
-    private readonly productRepository: ProductRepository,
-    private readonly loader: LoaderService
-  ) {
+  constructor(private readonly productRepository: ProductRepository) {
     super(initialState);
   }
 
   // Actions - public
   loadSupermarkets(selected: { [key: string]: boolean }, query?: string) {
     if (query) {
-      this.loader.setLoading(true);
       forkJoin([...this.loadExternalProductEffect(query, selected)])
         .pipe(
           tap(
@@ -56,7 +51,6 @@ export class ProductState extends SignalService<initialProductState> {
                 ...alcampoProducts,
                 ...gadisProducts,
               ]);
-              this.loader.setLoading(false);
             }
           )
         )
