--- src/app/app.config.ts
@@ -1,35 +1,0 @@
-import { ApplicationConfig, importProvidersFrom } from '@angular/core';
-import { provideRouter } from '@angular/router';
-
-import { routes } from './app.routes';
-import {
-  BrowserModule,
-  provideClientHydration,
-} from '@angular/platform-browser';
-import { provideHttpClient, withFetch } from '@angular/common/http';
-import { FormsModule } from '@angular/forms';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { provideToastr } from 'ngx-toastr';
-import { BrowserStorageService } from './shared/services/browser-storage.service';
-
-export const appConfig: ApplicationConfig = {
-  providers: [
-    provideHttpClient(withFetch()),
-    provideRouter(routes),
-    provideClientHydration(),
-    // {
-    //   provide: HTTP_INTERCEPTORS,
-    //   useClass: LoadingInterceptor,
-    //   multi: true,
-    // },
-    importProvidersFrom(
-      BrowserModule,
-
-      FormsModule,
-      BrowserAnimationsModule
-    ),
-    BrowserStorageService,
-
-    provideToastr(), // Toastr providers),
-  ],
-};
