+++ src/app/shared/components/add-button/add-button.component.ts
@@ -10,7 +10,6 @@ import { ToastrService } from 'ngx-toastr';
   styleUrl: './add-button.component.css',
   standalone: true,
   imports: [NgClass],
-  providers: [EventsStorageService],
 })
 export class AddButtonComponent {
   @Input() product!: ExternalProduct;
