+++ src/app/shared/components/navbar/navbar.component.css
@@ -34,10 +34,6 @@ a.active-link {
     box-shadow: 0px 0px 4px 0px #262626;
 }
 
-.icon-nav {
-    font-size: 14px;
-}
-
 /* @media (max-width: 1000px) {
     .bottom-bar {
         bottom: 0;
