+++ src/app/shared/services/events-storage.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
-import { LocalStorageService } from './local-storage.service';
-import { ExternalProduct } from '../interfaces/products.interface';
+import { ExternalProduct } from 'src/app/shared/interfaces/products.interface';
+import { LocalStorageService } from 'src/app/shared/services/local-storage.service';
 
-@Injectable({ providedIn: 'any' })
+@Injectable({ providedIn: 'root' })
 export class EventsStorageService extends LocalStorageService {
   override myLocalStorageTag = 'myListStorage';
   override storageType: 'ONE' | 'MULTIPLE' = 'MULTIPLE';
