+++ src/index.html
@@ -4,10 +4,6 @@
     <meta charset="utf-8" />
     <title>Super más barato</title>
     <base href="/" />
-    <meta
-      name="description"
-      content="Comparador de precios entre los supermercados más habituales"
-    />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta
       name="description"
